// src/app-react/components/dashboard-visualization/components/Histogram/types.ts
export type HistogramMode = "showBins" | "showLines" | "showLinesAndPoints";

export type StoredChartState = {
  mode: HistogramMode;
  scaleType: ScaleType;
  tickType: "center" | "bounds";
  selectedChannel: string;
  selectedIndex: number;
  tooltipsEnabled: boolean;
  showMinimap: boolean;
  zoomRange: NumberRange | null;
};

export type HistogramContextData = {
  chartData: ChartData;
  refreshActionsRef: RefObject<(() => void) | null>;
  setRef: (element: HTMLDivElement | null) => void;
  setMinimapRef: (element: HTMLDivElement | null) => void;
  setSelectedChannel: (channel: string) => void;
  setSelectedIndex: (selectedIndex: number) => void;
  visError: VIS_ERROR_TYPES | null;
  visMessage: string | null;
};

export type ScaleType = "linear" | "logarithmic";